refactor to reduce size of `agentic_system`
